Separate initialization object for buttons
Instead of relying on global variables for binding button click events,
make a second application class and pass the button elements, as
well as the map application instance, as options. Then bind the
events inside the constructor.

Use regular JavaScript instead of jQuery for event binding.
We still use it to request the JSON but we can replace this with
a fetch polyfill or just plain old XHR.