Stub out scale initialization
We want to use d3 scales to simplify calculating the choropleth
opacity.  Make a placeholder function to create our lookup table
of data property to scale.  Use this to set this.scales inside
the constructor.

Use a private property, _propertyToMap to store the changing
choropleth fill property instead of altering options.propertyToMap.
We should think of options as initial values that don't change and
store values that represent state, in a different way.