Use two yScales
Because the SVG origin is upper left, the logic for all heights and Y placements is inverted and unintuitive. Therefore, I'm using this other scale, with an inverted range, so that the bars can be drawn/placed using logic that makes sense to me and the scales can have a zero baseline at the bottom.