Forward & back
Add forward/back buttons to cycle through the pollution events one by one. This requires three new methods of the cafo map and a new class of buttons. the CafoMap.showProfile method is renamed and rejiggered to use an array index. That method also is used by the next() and back() methods, which do the wraparound math to make an infinite loop.